# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"EExchange": {
				"enum": [
					"COINGECKO_API_URL",
					"COINBASE_API_URL",
					"COINMARKETCAP",
					"UNISWAP"
				],
				"type": "string"
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "apiKey",
				"name": "Authorization",
				"scheme": "bearer",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "api-flash",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/contract/contact": {
			"post": {
				"operationId": "Contact",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {},
										"token": {
											"type": "string"
										},
										"data": {
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Verificar contrato.",
				"tags": [
					"Contract"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/contract/all": {
			"get": {
				"operationId": "All",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {},
										"token": {
											"type": "string"
										},
										"data": {
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Verificar contrato.",
				"tags": [
					"Contract"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "type",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/EExchange"
						}
					},
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "init",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/contract/dinamicContract": {
			"post": {
				"operationId": "DinamicContract",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {},
										"token": {
											"type": "string"
										},
										"data": {
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Contract"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"amount": {
										"type": "string"
									}
								},
								"required": [
									"amount"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}